Move IListDataSource impl. to ListStoreBase
