+++ Xwt/Xwt.csproj
@@ -326,6 +326,7 @@
     <Compile Include="Xwt.Backends\IColorPickerBackend.cs" />
     <Compile Include="Xwt\XwtSynchronizationContext.cs" />
     <Compile Include="Xwt\TreeViewNavigator.cs" />
+    <Compile Include="Xwt\ListStoreBase.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup />
